+++ src/main/java/absolutelyaya/ultracraft/block/FakeLeavesBlock.java
@@ -1,0 +1,35 @@
+package absolutelyaya.ultracraft.block;
+
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.state.StateManager;
+import net.minecraft.state.property.BooleanProperty;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.util.shape.VoxelShapes;
+import net.minecraft.world.BlockView;
+
+public class FakeLeavesBlock extends Block
+{
+	public static BooleanProperty GLITCH = BooleanProperty.of("glitch");
+	
+	public FakeLeavesBlock(Settings settings)
+	{
+		super(settings);
+		setDefaultState(getDefaultState().with(GLITCH, false));
+	}
+	
+	@Override
+	protected void appendProperties(StateManager.Builder<Block, BlockState> builder)
+	{
+		super.appendProperties(builder);
+		builder.add(GLITCH);
+	}
+	
+	@Override
+	public VoxelShape getCollisionShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context)
+	{
+		return VoxelShapes.empty();
+	}
+}